requirements.txt
aiohttp==3.7.4.post0
aniso8601==9.0.1
async-timeout==3.0.1
attrs==21.2.0
certifi==2020.12.5
chardet==4.0.0
cheroot==8.5.2
click==7.1.2
colorama==0.4.4
dateparser==1.0.0
Flask==1.1.2
Flask-RESTful==0.3.8
idna==2.10
itsdangerous==1.1.0
jaraco.functools==3.3.0
Jinja2==2.11.3
MarkupSafe==1.1.1
more-itertools==8.7.0
multidict==5.1.0
numpy==1.20.3
pathlib==1.0.1
psutil==5.8.0
python-binance==1.0.7
python-dateutil==2.8.1
python-dotenv==0.17.1
pytz==2021.1
regex==2021.4.4
requests==2.25.1
six==1.16.0
typing-extensions==3.10.0.0
tzlocal==2.1
ujson==4.0.2
unicorn-binance-websocket-api==1.30.0
unicorn-fy==0.11.0
urllib3==1.26.4
websocket-client==0.59.0
websockets==8.1
Werkzeug==1.0.1
yarl==1.6.3
